// frontend/src/app/api/model/recurringScheduleTemplateDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RecurringScheduleTemplateDto { 
    rrule?: string;
    startDateTime?: string;
    endDateTime?: string;
}